archive/pyproject.toml
[project]
name = "archive"
version = "0.1.1"
description = "Command line tool for building data archives"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.0.0",
    "colorama>=0.4.0",
    "pyyaml>=6.0.2",
    "types-pyyaml>=6.0.12.20250516",
]

[project.scripts]
archive = "archive.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
    "types-colorama>=0.4.15.20240311",
    "mypy>=1.15.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "ISC", "PIE", "T20", "Q"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=archive --cov-report=term-missing"

